Use a template to create/update the .gitconfig file
Introduced a template _gitconfig.tmpl_ that is used to create and update
the _.gitconfig_. This allows setting the user name and email without
committing them to git. The _.gitconfig_ is always updated when opening
a new shell and has not recently been modified. If it was modified, it
is copied back to the template file so that changes are not overwritten
and can be updated in the VCS.