Add git update-ref function for upstream branches
Added convenience function `guru` that allows to update a branch to
the commit ref of its upstream branch. The branch to be updated has
to be provided as first parameter (and only) parameter. If none is
provided, the `master` branch is updated.

The previously defined `gur-*` aliases have been removed.